CRUD

+++ src/main/java/com/p5/adoptionsapi/repository/animal/Animal.java
@@ -2,14 +2,6 @@ package com.p5.adoptionsapi.repository.animal;
 
 import javax.persistence.*;
 
-//Option 1
-//@MappedSuperclass
-
-//Option 2
-//@Entity
-//@Inheritance(strategy = InheritanceType.SINGLE_TABLE)
-////Only if we use SINGLE_TABLE
-//@DiscriminatorColumn(name = "type", discriminatorType = DiscriminatorType.STRING)
 
 @Entity
 @Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
@@ -17,11 +9,11 @@ public class Animal {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private Integer id;
     private String name;
     private String photo;
 
-    public Animal(Long id, String name, String photo) {
+    public Animal(Integer id, String name, String photo) {
         this.id = id;
         this.name = name;
         this.photo = photo;
@@ -30,11 +22,11 @@ public class Animal {
     public Animal() {
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public Animal setId(Long id) {
+    public Animal setId(Integer id) {
         this.id = id;
         return this;
     }
